test files for cobertura bad metrics
