+++ test/tmp_workspace/cobertura-sample/.gitignore
@@ -1,12 +1,13 @@
-cobertura-sample/target/
-*.class
-*.ser
-cobertura-sample/target/site/cobertura/
-cobertura-sample/target/generated-sources/
-cobertura-sample/target/generated-test-sources/
-cobertura-sample/target/generated-classes/
-cobertura-sample/target/maven-status/
-cobertura-sample/target/surefire-reports/
+bad-metrics-project/target/
+bad-metrics-project/target/*.class
+bad-metrics-project/target/*.ser
+bad-metrics-project/target/site/
+bad-metrics-project/target/surefire-reports/
+bad-metrics-project/target/maven-status/
+bad-metrics-project/target/cobertura/
+bad-metrics-project/target/generated-sources/
+bad-metrics-project/target/generated-test-sources/
+bad-metrics-project/target/generated-classes/
 *.log
 *.tmp
 *.swp
@@ -14,3 +15,9 @@ cobertura-sample/target/surefire-reports/
 *.iml
 *.project
 *.classpath
+**/.idea/
+**/*.iml
+**/.classpath
+**/.project
+*.png
+
