+++ test/tmp_workspace/cobertura-sample/bad-metrics-project/src/main/java/com/example/package1/Calculator.java
@@ -1,0 +1,20 @@
+package com.example.package1;
+
+public class Calculator {
+    public int add(int a, int b) {
+        return a + b;
+    }
+
+    public int subtract(int a, int b) {
+        return a - b;
+    }
+
+    public int multiply(int a, int b) {
+        return a * b;
+    }
+
+    public int divide(int a, int b) {
+        if (b == 0) throw new IllegalArgumentException("Cannot divide by zero");
+        return a / b;
+    }
+}
