+++ test/tmp_workspace/cobertura-sample/bad-metrics-project/src/main/java/com/example/package2/Multiplier.java
@@ -1,0 +1,11 @@
+package com.example.package2;
+
+public class Multiplier {
+    public int multiply(int a, int b) {
+        int result = 0;
+        for (int i = 0; i < b; i++) {
+            result += a;
+        }
+        return result;
+    }
+}
