+++ test/tmp_workspace/cobertura-sample/bad-metrics-project/src/test/java/com/example/package1/CalculatorTest.java
@@ -1,0 +1,12 @@
+package com.example.package1;
+
+import org.junit.Test;
+import static org.junit.Assert.*;
+
+public class CalculatorTest {
+    @Test
+    public void testAdd() {
+        Calculator calculator = new Calculator();
+        assertEquals(5, calculator.add(2, 3));
+    }
+}
