--- test/tmp_workspace/cobertura-sample/src/main/java/com/example/package1/Calculator.java
@@ -1,7 +1,0 @@
-package com.example.package1;
-
-public class Calculator {
-    public int add(int a, int b) {
-        return a + b;
-    }
-}
