--- test/tmp_workspace/cobertura-sample/src/main/java/com/example/package2/Multiplier.java
@@ -1,7 +1,0 @@
-package com.example.package2;
-
-public class Multiplier {
-    public int multiply(int a, int b) {
-        return a * b;
-    }
-}
