--- test/tmp_workspace/cobertura-sample/src/main/java/com/example/package3/Divider.java
@@ -1,8 +1,0 @@
-package com.example.package3;
-
-public class Divider {
-    public int divide(int a, int b) {
-        if (b == 0) throw new ArithmeticException("Cannot divide by zero");
-        return a / b;
-    }
-}
