--- test/tmp_workspace/cobertura-sample/src/test/java/com/example/package1/CalculatorTest.java
@@ -1,12 +1,0 @@
-package com.example.package1;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class CalculatorTest {
-    @Test
-    public void testAdd() {
-        Calculator calculator = new Calculator();
-        assertEquals(7, calculator.add(3, 4));
-    }
-}
