--- test/tmp_workspace/cobertura-sample/src/test/java/com/example/package2/MultiplierTest.java
@@ -1,12 +1,0 @@
-package com.example.package2;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class MultiplierTest {
-    @Test
-    public void testMultiply() {
-        Multiplier multiplier = new Multiplier();
-        assertEquals(12, multiplier.multiply(3, 4));
-    }
-}
